refactor(filter): rename request_validate module to validate
The request_validate module name was redundant since it already lives
under the responses directory. Also renames the inner validate.rs to
rules.rs to avoid clippy::module_inception.
