+++ filter/src/builtins/http/ai/openai/responses/validate/rules.rs
